refactor: Update PizzaIngredient table columns
The commit updates the columns of the `PizzaIngredient` table in the `config.ts` file. The `pizza_id` column now has a reference to the `id` column of the `Pizza` table, and the `ingredient_id` column now has a reference to the `id` column of the `Ingredient` table. This change ensures the integrity of the relationships between the tables.